tools/discord-issue-bot/requirements.txt
discord.py>=2.3,<3.0
PyGithub>=2.1,<3.0
python-dotenv>=1.0,<2.0
